refactor : 검증 로직 수정 및 패키지 변경
