+++ src/test/java/roomescape/MissionStepTest.java
@@ -65,7 +65,7 @@ public class MissionStepTest {
         RestAssured.given().log().all()
                 .when().delete("/reservations/1")
                 .then().log().all()
-                .statusCode(400);
+                .statusCode(404);
     }
 
 }
